Use pip instead of apt-get for package installation
Otherwise the `ansible` that is found in the PATH isn't the one from
the debian package, but an ansible installation that is pre-installed
in the runner which brings its own virtualenv and thus doesn't find
the system-installed boto3.

One could probably call ansible with an absolute path to amend this.